Add warnings and optimize flag to compiler and monitor_speed to platformio.ini

+++ src/modules/pwnagotchi/pwngrid.h
@@ -26,7 +26,7 @@ typedef struct {
   int uptime;
   String version;
   signed int rssi;
-  int last_ping;
+  unsigned long last_ping;
   bool gone;
 } pwngrid_peer;
